correct t7 switch int and float

+++ python_variable_annotations/7-to_kv.py
@@ -3,7 +3,7 @@
 from typing import List, Union, Tuple
 
 
-def to_kv(k: str, v: Union[float, int]) -> Tuple[str, float]:
+def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """takes a string k and an int OR float v as arguments and returns a
     tuple."""
     return k, float(v * v)
